:sparkles: 결제 기능 추가 (페이팔, 디파짓 결제) + 가격 단위 변경

+++ src/main/java/com/backend/supermeproject/member/dto/MyPageDto.java
@@ -2,6 +2,8 @@ package com.backend.supermeproject.member.dto;
 
 import com.backend.supermeproject.global.role.Gender;
 
+import java.math.BigDecimal;
+
 public record MyPageDto(
         String profileImage,
         String name,
@@ -12,7 +14,7 @@ public record MyPageDto(
         String city,
         String postcode,
         Gender gender,
-        Long deposit
+        BigDecimal deposit
         ) {
